feat: added initializeCreateListing to createListing instead of home.js
